Atualizações em routes, validator, migration e controller Concertando comentario no routes.js Melhorando criação de manifestação Atualizando o validator para uma classe que contem o save e o update Migration: secretary_id pode ser null e user_id não pode
